core: prevent simultaneous deployment of a vm

+++ core/src/main/java/io/github/floto/core/jobs/RedeployVmJob.java
@@ -8,6 +8,7 @@ import io.github.floto.dsl.model.VmConfiguration;
 
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,6 +16,7 @@ import org.slf4j.LoggerFactory;
 public class RedeployVmJob extends HypervisorJob<Void> {
     private Logger log = LoggerFactory.getLogger(RedeployVmJob.class);
     private FlotoService flotoService;
+	static CopyOnWriteArrayList<String> deployVmList = new CopyOnWriteArrayList<>();
 
     public RedeployVmJob(FlotoService flotoService, String vmName) {
         super(flotoService.getManifest(), vmName);
@@ -23,36 +25,47 @@ public class RedeployVmJob extends HypervisorJob<Void> {
 
     @Override
     public Void execute() throws Exception {
-        VmConfiguration vmConfiguration = host.vmConfiguration;
-        VmDescription vmDescription = new VmDescription();
-        vmDescription.vmName = host.name;
-        vmDescription.numberOfCores = vmConfiguration.numberOfCores;
-        vmDescription.memoryInMB = vmConfiguration.memoryInMB;
-        vmDescription.vmNetworks = new ArrayList<>(vmConfiguration.networks);
-        for (DiskDescription diskDesc : vmConfiguration.disks) {
-        	Disk disk = new Disk(vmDescription);
-        	disk.sizeInGB = diskDesc.sizeInGB;
-        	disk.datastore = diskDesc.datastore;
-        	disk.slot = diskDesc.slot;
-        	disk.mountpoint = diskDesc.mountpoint;
-        	vmDescription.disks.add(disk);
-        }
+		if (!deployVmList.addIfAbsent(host.name)) {
+			throw new RuntimeException(host.name + " already in deployment");
+		}
 
-        log.info("Removing old VM");
-        hypervisorService.stopVm(host.name);
-        hypervisorService.deleteVm(host.name);
+		try {
 
-        log.info("Deploying VM {}", host.name);
-        URL ovaUrl = new URL(vmConfiguration.ovaUrl);
-        hypervisorService.deployVm(ovaUrl, vmDescription);
+			VmConfiguration vmConfiguration = host.vmConfiguration;
+			VmDescription vmDescription = new VmDescription();
+			vmDescription.vmName = host.name;
+			vmDescription.numberOfCores = vmConfiguration.numberOfCores;
+			vmDescription.memoryInMB = vmConfiguration.memoryInMB;
+			vmDescription.vmNetworks = new ArrayList<>(vmConfiguration.networks);
+			for (DiskDescription diskDesc : vmConfiguration.disks) {
+				Disk disk = new Disk(vmDescription);
+				disk.sizeInGB = diskDesc.sizeInGB;
+				disk.datastore = diskDesc.datastore;
+				disk.slot = diskDesc.slot;
+				disk.mountpoint = diskDesc.mountpoint;
+				vmDescription.disks.add(disk);
+			}
 
-        log.info("Starting VM {}", host.name);
-        hypervisorService.startVm(host.name);
+			log.info("Removing old VM");
+			hypervisorService.stopVm(host.name);
+			hypervisorService.deleteVm(host.name);
 
-        runPostDeploy();
+			log.info("Deploying VM {}", host.name);
+			URL ovaUrl = new URL(vmConfiguration.ovaUrl);
+			hypervisorService.deployVm(ovaUrl, vmDescription);
 
-        log.info("Post-deploy completed on host {}", host.name);
-        return null;
+			log.info("Starting VM {}", host.name);
+			hypervisorService.startVm(host.name);
+
+			runPostDeploy();
+
+			log.info("Post-deploy completed on host {}", host.name);
+
+		} finally {
+			deployVmList.remove(host.name);
+		}
+
+		return null;
     }
 
     private void runPostDeploy() {
